NVC-11
- Alterações feitas no tipo do id dos enums Status e Estado;
- Criada uma classe abstrata para extender o id;
- Criado um Mapper para EmpresaDTO;
- Criada a Controller de Empresa;
- Criada a Controller de Entidade;
- Criada a Service de Empresa;
- Criada a Service de Entidade;
- Criado um Mapper para EntidadeDTO.

+++ src/main/java/br/com/dev/ecommerce/admin/entidade/enums/Status.java
@@ -2,24 +2,22 @@ package br.com.dev.ecommerce.admin.entidade.enums;
 
 public enum Status {
 
-	INATIVO(0, "Inativo"), 
-	ATIVO(1, "Ativo"), 
-	EXCLUIDO(2, "Excluido");
-	
-	private int id;
-	
+	INATIVO(0L, "Inativo"), ATIVO(1L, "Ativo"), EXCLUIDO(2L, "Excluido");
+
+	private Long id;
+
 	private String descricao;
-	
-	private Status(int id, String descricao) {
+
+	private Status(Long id, String descricao) {
 		this.id = id;
 		this.descricao = descricao;
 	}
 
-	public int getId() {
+	public Long getId() {
 		return id;
 	}
 
-	public void setId(int id) {
+	public void setId(Long id) {
 		this.id = id;
 	}
